trunk/documentation/evalactipol_doc/SimpleTest/WebTester/TestRadioFields.html

<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN">
<html>
<head>
	<title>Docs For Class TestRadioFields</title>
	<link rel="stylesheet" type="text/css" id="layout" href="../../media/layout.css" media="screen">
	<link rel="stylesheet" type="text/css" href="../../media/style.css" media="all">
	<link rel="stylesheet" type="text/css" href="../../media/print.css" media="print">
</head>

<body>
<div id="header">
	<div id="navLinks">
        
                                                                                                
                            	        [ <a href="../../classtrees_SimpleTest.html">Class Tree: SimpleTest</a> ]
            [ <a href="../../elementindex_SimpleTest.html">Index: SimpleTest</a> ]
        		[ <a href="../../elementindex.html">All elements</a> ]		
	</div>
	<div id="packagePosition">
		<div id="packageTitle2">SimpleTest</div>
		<div id="packageTitle">SimpleTest</div>
		<div id="elementPath">WebTester &middot; </div>
	</div>
</div>

<div id="nav" class="small">
	<div id="packages">
		Packages:
					<p><a href="../../li_default.html">default</a></p>
					<p><a href="../../li_Cache_Lite.html">Cache_Lite</a></p>
					<p><a href="../../li_Evalactipol_libraries.html">Evalactipol_libraries</a></p>
					<p><a href="../../li_SimpleTest.html">SimpleTest</a></p>
			</div>

				<div id="index">
			<div id="files">
								Files:<br>
								<a href="../../SimpleTest/_simpletest---autorun.php.html">		autorun.php
		</a><br>
			<a href="../../SimpleTest/_simpletest---compatibility.php.html">		compatibility.php
		</a><br>
		subpackage <b>Eclipse</b><br>			<a href="../../SimpleTest/Eclipse/_simpletest---eclipse.php.html">		eclipse.php
		</a><br>
		subpackage <b>Extensions</b><br>			<a href="../../SimpleTest/Extensions/_simpletest---extensions---pear_test_case.php.html">		pear_test_case.php
		</a><br>
			<a href="../../SimpleTest/Extensions/_simpletest---extensions---phpunit_test_case.php.html">		phpunit_test_case.php
		</a><br>
		subpackage <b>MockObjects</b><br>			<a href="../../SimpleTest/MockObjects/_simpletest---mock_objects.php.html">		mock_objects.php
		</a><br>
			<a href="../../SimpleTest/MockObjects/_simpletest---parser.php.html">		parser.php
		</a><br>
			<a href="../../SimpleTest/MockObjects/_simpletest---socket.php.html">		socket.php
		</a><br>
		subpackage <b>UnitTester</b><br>			<a href="../../SimpleTest/UnitTester/_simpletest---collector.php.html">		collector.php
		</a><br>
			<a href="../../SimpleTest/UnitTester/_simpletest---default_reporter.php.html">		default_reporter.php
		</a><br>
			<a href="../../SimpleTest/UnitTester/_simpletest---detached.php.html">		detached.php
		</a><br>
			<a href="../../SimpleTest/UnitTester/_simpletest---dumper.php.html">		dumper.php
		</a><br>
			<a href="../../SimpleTest/UnitTester/_simpletest---errors.php.html">		errors.php
		</a><br>
			<a href="../../SimpleTest/UnitTester/_simpletest---exceptions.php.html">		exceptions.php
		</a><br>
			<a href="../../SimpleTest/UnitTester/_simpletest---expectation.php.html">		expectation.php
		</a><br>
			<a href="../../SimpleTest/UnitTester/_simpletest---invoker.php.html">		invoker.php
		</a><br>
			<a href="../../SimpleTest/UnitTester/_simpletest---reflection_php4.php.html">		reflection_php4.php
		</a><br>
			<a href="../../SimpleTest/UnitTester/_simpletest---reflection_php5.php.html">		reflection_php5.php
		</a><br>
			<a href="../../SimpleTest/UnitTester/_simpletest---remote.php.html">		remote.php
		</a><br>
			<a href="../../SimpleTest/UnitTester/_simpletest---reporter.php.html">		reporter.php
		</a><br>
			<a href="../../SimpleTest/UnitTester/_simpletest---scorer.php.html">		scorer.php
		</a><br>
			<a href="../../SimpleTest/UnitTester/_simpletest---shell_tester.php.html">		shell_tester.php
		</a><br>
			<a href="../../SimpleTest/UnitTester/_simpletest---simpletest.php.html">		simpletest.php
		</a><br>
			<a href="../../SimpleTest/UnitTester/_simpletest---test_case.php.html">		test_case.php
		</a><br>
			<a href="../../SimpleTest/UnitTester/_simpletest---unit_tester.php.html">		unit_tester.php
		</a><br>
			<a href="../../SimpleTest/UnitTester/_simpletest---xml.php.html">		xml.php
		</a><br>
		subpackage <b>WebTester</b><br>			<a href="../../SimpleTest/WebTester/_simpletest---authentication.php.html">		authentication.php
		</a><br>
			<a href="../../SimpleTest/WebTester/_simpletest---browser.php.html">		browser.php
		</a><br>
			<a href="../../SimpleTest/WebTester/_simpletest---cookies.php.html">		cookies.php
		</a><br>
			<a href="../../SimpleTest/WebTester/_simpletest---encoding.php.html">		encoding.php
		</a><br>
			<a href="../../SimpleTest/WebTester/_simpletest---form.php.html">		form.php
		</a><br>
			<a href="../../SimpleTest/WebTester/_simpletest---frames.php.html">		frames.php
		</a><br>
			<a href="../../SimpleTest/WebTester/_simpletest---http.php.html">		http.php
		</a><br>
			<a href="../../SimpleTest/WebTester/_simpletest---page.php.html">		page.php
		</a><br>
			<a href="../../SimpleTest/WebTester/_simpletest---selector.php.html">		selector.php
		</a><br>
			<a href="../../SimpleTest/WebTester/_simpletest---tag.php.html">		tag.php
		</a><br>
			<a href="../../SimpleTest/WebTester/_simpletest---url.php.html">		url.php
		</a><br>
			<a href="../../SimpleTest/WebTester/_simpletest---user_agent.php.html">		user_agent.php
		</a><br>
			<a href="../../SimpleTest/WebTester/_simpletest---web_tester.php.html">		web_tester.php
		</a><br>
				</div>
			<div id="interfaces">
							</div>
			<div id="classes">
				Classes:<br>
					<b>WebTester</b><br>			<a href="../../SimpleTest/WebTester/FieldExpectation.html">		FieldExpectation
		</a><br>
			<a href="../../SimpleTest/WebTester/HttpHeaderExpectation.html">		HttpHeaderExpectation
		</a><br>
			<a href="../../SimpleTest/WebTester/LiveTestOfForms.html">		LiveTestOfForms
		</a><br>
			<a href="../../SimpleTest/WebTester/MultipleSelectionTag.html">		MultipleSelectionTag
		</a><br>
			<a href="../../SimpleTest/WebTester/NoHttpHeaderExpectation.html">		NoHttpHeaderExpectation
		</a><br>
			<a href="../../SimpleTest/WebTester/ParallelRegex.html">		ParallelRegex
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleAnchorTag.html">		SimpleAnchorTag
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleAttachment.html">		SimpleAttachment
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleAuthenticator.html">		SimpleAuthenticator
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleBaseTag.html">		SimpleBaseTag
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleBrowser.html">		SimpleBrowser
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleBrowserHistory.html">		SimpleBrowserHistory
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleButtonTag.html">		SimpleButtonTag
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleById.html">		SimpleById
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleByLabel.html">		SimpleByLabel
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleByLabelOrName.html">		SimpleByLabelOrName
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleByName.html">		SimpleByName
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleCheckboxGroup.html">		SimpleCheckboxGroup
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleCheckboxTag.html">		SimpleCheckboxTag
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleCookie.html">		SimpleCookie
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleCookieJar.html">		SimpleCookieJar
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleEncodedPair.html">		SimpleEncodedPair
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleEncoding.html">		SimpleEncoding
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleForm.html">		SimpleForm
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleFormTag.html">		SimpleFormTag
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleFrameset.html">		SimpleFrameset
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleFrameTag.html">		SimpleFrameTag
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleGetEncoding.html">		SimpleGetEncoding
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleHeadEncoding.html">		SimpleHeadEncoding
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleHtmlLexer.html">		SimpleHtmlLexer
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleHtmlSaxParser.html">		SimpleHtmlSaxParser
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleHttpHeaders.html">		SimpleHttpHeaders
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleHttpRequest.html">		SimpleHttpRequest
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleHttpResponse.html">		SimpleHttpResponse
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleImageSubmitTag.html">		SimpleImageSubmitTag
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleLabelTag.html">		SimpleLabelTag
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleLexer.html">		SimpleLexer
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleMultipartEncoding.html">		SimpleMultipartEncoding
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleOptionTag.html">		SimpleOptionTag
		</a><br>
			<a href="../../SimpleTest/WebTester/SimplePage.html">		SimplePage
		</a><br>
			<a href="../../SimpleTest/WebTester/SimplePageBuilder.html">		SimplePageBuilder
		</a><br>
			<a href="../../SimpleTest/WebTester/SimplePostEncoding.html">		SimplePostEncoding
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleProxyRoute.html">		SimpleProxyRoute
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleRadioButtonTag.html">		SimpleRadioButtonTag
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleRadioGroup.html">		SimpleRadioGroup
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleRealm.html">		SimpleRealm
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleRoute.html">		SimpleRoute
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleSaxListener.html">		SimpleSaxListener
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleSecureSocket.html">		SimpleSecureSocket
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleSelectionTag.html">		SimpleSelectionTag
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleSocket.html">		SimpleSocket
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleStateStack.html">		SimpleStateStack
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleStickyError.html">		SimpleStickyError
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleSubmitTag.html">		SimpleSubmitTag
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleTag.html">		SimpleTag
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleTagBuilder.html">		SimpleTagBuilder
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleTagGroup.html">		SimpleTagGroup
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleTestAcceptanceTest.html">		SimpleTestAcceptanceTest
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleTextAreaTag.html">		SimpleTextAreaTag
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleTextTag.html">		SimpleTextTag
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleTitleTag.html">		SimpleTitleTag
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleUploadTag.html">		SimpleUploadTag
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleUrl.html">		SimpleUrl
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleUserAgent.html">		SimpleUserAgent
		</a><br>
			<a href="../../SimpleTest/WebTester/SimpleWidget.html">		SimpleWidget
		</a><br>
			<a href="../../SimpleTest/WebTester/TestOfFileUploads.html">		TestOfFileUploads
		</a><br>
			<a href="../../SimpleTest/WebTester/TestOfFrameAuthentication.html">		TestOfFrameAuthentication
		</a><br>
			<a href="../../SimpleTest/WebTester/TestOfGenericAssertionsInWebTester.html">		TestOfGenericAssertionsInWebTester
		</a><br>
			<a href="../../SimpleTest/WebTester/TestOfLinkFollowing.html">		TestOfLinkFollowing
		</a><br>
			<a href="../../SimpleTest/WebTester/TestOfLiveAuthentication.html">		TestOfLiveAuthentication
		</a><br>
			<a href="../../SimpleTest/WebTester/TestOfLiveCookies.html">		TestOfLiveCookies
		</a><br>
			<a href="../../SimpleTest/WebTester/TestOfLiveFetching.html">		TestOfLiveFetching
		</a><br>
			<a href="../../SimpleTest/WebTester/TestOfLiveFrontControllerEmulation.html">		TestOfLiveFrontControllerEmulation
		</a><br>
			<a href="../../SimpleTest/WebTester/TestOfLiveHeaders.html">		TestOfLiveHeaders
		</a><br>
			<a href="../../SimpleTest/WebTester/TestOfLiveHistoryNavigation.html">		TestOfLiveHistoryNavigation
		</a><br>
			<a href="../../SimpleTest/WebTester/TestOfLiveMultiValueWidgets.html">		TestOfLiveMultiValueWidgets
		</a><br>
			<a href="../../SimpleTest/WebTester/TestOfLivePageLinkingWithMinimalLinks.html">		TestOfLivePageLinkingWithMinimalLinks
		</a><br>
			<a href="../../SimpleTest/WebTester/TestOfLiveRedirects.html">		TestOfLiveRedirects
		</a><br>
			<a href="../../SimpleTest/WebTester/TestOfLoadingFrames.html">		TestOfLoadingFrames
		</a><br>
			<a href="../../SimpleTest/WebTester/TestOfNestedFrames.html">		TestOfNestedFrames
		</a><br>
			<a href="../../SimpleTest/WebTester/TestRadioFields.html">		TestRadioFields
		</a><br>
			<a href="../../SimpleTest/WebTester/WebTestCase.html">		WebTestCase
		</a><br>
		<b>UnitTester</b><br>			<a href="../../SimpleTest/UnitTester/AllTests.html">		AllTests
		</a><br>
			<a href="../../SimpleTest/UnitTester/BadGroupTest.html">		BadGroupTest
		</a><br>
			<a href="../../SimpleTest/UnitTester/BadTestCases.html">		BadTestCases
		</a><br>
			<a href="../../SimpleTest/UnitTester/BadTestSuite.html">		BadTestSuite
		</a><br>
			<a href="../../SimpleTest/UnitTester/DefaultReporter.html">		DefaultReporter
		</a><br>
			<a href="../../SimpleTest/UnitTester/DetachedTestCase.html">		DetachedTestCase
		</a><br>
			<a href="../../SimpleTest/UnitTester/EqualExpectation.html">		EqualExpectation
		</a><br>
			<a href="../../SimpleTest/UnitTester/ExceptionExpectation.html">		ExceptionExpectation
		</a><br>
			<a href="../../SimpleTest/UnitTester/FailingUnitTestCaseOutput.html">		FailingUnitTestCaseOutput
		</a><br>
			<a href="../../SimpleTest/UnitTester/GroupTest.html">		GroupTest
		</a><br>
			<a href="../../SimpleTest/UnitTester/HtmlReporter.html">		HtmlReporter
		</a><br>
			<a href="../../SimpleTest/UnitTester/IdenticalExpectation.html">		IdenticalExpectation
		</a><br>
			<a href="../../SimpleTest/UnitTester/IsAExpectation.html">		IsAExpectation
		</a><br>
			<a href="../../SimpleTest/UnitTester/LikeExpectation.html">		LikeExpectation
		</a><br>
			<a href="../../SimpleTest/UnitTester/LiveHttpTestCase.html">		LiveHttpTestCase
		</a><br>
			<a href="../../SimpleTest/UnitTester/LoadIfIncludedTestCase.html">		LoadIfIncludedTestCase
		</a><br>
			<a href="../../SimpleTest/UnitTester/MethodExistsExpectation.html">		MethodExistsExpectation
		</a><br>
			<a href="../../SimpleTest/UnitTester/MultipleReporter.html">		MultipleReporter
		</a><br>
			<a href="../../SimpleTest/UnitTester/NestingCaseTag.html">		NestingCaseTag
		</a><br>
			<a href="../../SimpleTest/UnitTester/NestingGroupTag.html">		NestingGroupTag
		</a><br>
			<a href="../../SimpleTest/UnitTester/NestingMethodTag.html">		NestingMethodTag
		</a><br>
			<a href="../../SimpleTest/UnitTester/NestingXmlTag.html">		NestingXmlTag
		</a><br>
			<a href="../../SimpleTest/UnitTester/NoPatternExpectation.html">		NoPatternExpectation
		</a><br>
			<a href="../../SimpleTest/UnitTester/NoSkipsReporter.html">		NoSkipsReporter
		</a><br>
			<a href="../../SimpleTest/UnitTester/NotAExpectation.html">		NotAExpectation
		</a><br>
			<a href="../../SimpleTest/UnitTester/NotEqualExpectation.html">		NotEqualExpectation
		</a><br>
			<a href="../../SimpleTest/UnitTester/NoTextExpectation.html">		NoTextExpectation
		</a><br>
			<a href="../../SimpleTest/UnitTester/NotIdenticalExpectation.html">		NotIdenticalExpectation
		</a><br>
			<a href="../../SimpleTest/UnitTester/OutsideMarginExpectation.html">		OutsideMarginExpectation
		</a><br>
			<a href="../../SimpleTest/UnitTester/PassesAsWellReporter.html">		PassesAsWellReporter
		</a><br>
			<a href="../../SimpleTest/UnitTester/PassingUnitTestCaseOutput.html">		PassingUnitTestCaseOutput
		</a><br>
			<a href="../../SimpleTest/UnitTester/PathEqualExpectation.html">		PathEqualExpectation
		</a><br>
			<a href="../../SimpleTest/UnitTester/PatternExpectation.html">		PatternExpectation
		</a><br>
			<a href="../../SimpleTest/UnitTester/ReferenceExpectation.html">		ReferenceExpectation
		</a><br>
			<a href="../../SimpleTest/UnitTester/RemoteTestCase.html">		RemoteTestCase
		</a><br>
			<a href="../../SimpleTest/UnitTester/SelectiveReporter.html">		SelectiveReporter
		</a><br>
			<a href="../../SimpleTest/UnitTester/ShellTestCase.html">		ShellTestCase
		</a><br>
			<a href="../../SimpleTest/UnitTester/ShouldNeverBeRun.html">		ShouldNeverBeRun
		</a><br>
			<a href="../../SimpleTest/UnitTester/ShouldNeverBeRunEither.html">		ShouldNeverBeRunEither
		</a><br>
			<a href="../../SimpleTest/UnitTester/SimpleCollector.html">		SimpleCollector
		</a><br>
			<a href="../../SimpleTest/UnitTester/SimpleCommandLineParser.html">		SimpleCommandLineParser
		</a><br>
			<a href="../../SimpleTest/UnitTester/SimpleDumper.html">		SimpleDumper
		</a><br>
			<a href="../../SimpleTest/UnitTester/SimpleErrorQueue.html">		SimpleErrorQueue
		</a><br>
			<a href="../../SimpleTest/UnitTester/SimpleErrorTrappingInvoker.html">		SimpleErrorTrappingInvoker
		</a><br>
			<a href="../../SimpleTest/UnitTester/SimpleExceptionTrap.html">		SimpleExceptionTrap
		</a><br>
			<a href="../../SimpleTest/UnitTester/SimpleExceptionTrappingInvoker.html">		SimpleExceptionTrappingInvoker
		</a><br>
			<a href="../../SimpleTest/UnitTester/SimpleExpectation.html">		SimpleExpectation
		</a><br>
			<a href="../../SimpleTest/UnitTester/SimpleFileLoader.html">		SimpleFileLoader
		</a><br>
			<a href="../../SimpleTest/UnitTester/SimpleInvoker.html">		SimpleInvoker
		</a><br>
			<a href="../../SimpleTest/UnitTester/SimpleInvokerDecorator.html">		SimpleInvokerDecorator
		</a><br>
			<a href="../../SimpleTest/UnitTester/SimplePatternCollector.html">		SimplePatternCollector
		</a><br>
			<a href="../../SimpleTest/UnitTester/SimpleReflection.html">		SimpleReflection
		</a><br>
			<a href="../../SimpleTest/UnitTester/SimpleReporter.html">		SimpleReporter
		</a><br>
			<a href="../../SimpleTest/UnitTester/SimpleReporterDecorator.html">		SimpleReporterDecorator
		</a><br>
			<a href="../../SimpleTest/UnitTester/SimpleScorer.html">		SimpleScorer
		</a><br>
			<a href="../../SimpleTest/UnitTester/SimpleShell.html">		SimpleShell
		</a><br>
			<a href="../../SimpleTest/UnitTester/SimpleStackTrace.html">		SimpleStackTrace
		</a><br>
			<a href="../../SimpleTest/UnitTester/SimpleTest.html">		SimpleTest
		</a><br>
			<a href="../../SimpleTest/UnitTester/SimpleTestCase.html">		SimpleTestCase
		</a><br>
			<a href="../../SimpleTest/UnitTester/SimpleTestContext.html">		SimpleTestContext
		</a><br>
			<a href="../../SimpleTest/UnitTester/SimpleTestOptions.html">		SimpleTestOptions
		</a><br>
			<a href="../../SimpleTest/UnitTester/SimpleTestXmlParser.html">		SimpleTestXmlParser
		</a><br>
			<a href="../../SimpleTest/UnitTester/test1.html">		test1
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestCaseWithParseError.html">		TestCaseWithParseError
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestDoxReporter.html">		TestDoxReporter
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfAbsoluteUrls.html">		TestOfAbsoluteUrls
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfAbstractsWithAbstractMethods.html">		TestOfAbstractsWithAbstractMethods
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfAdditionalHeaders.html">		TestOfAdditionalHeaders
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfAnythingExpectation.html">		TestOfAnythingExpectation
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfAuthenticator.html">		TestOfAuthenticator
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfAuthorisation.html">		TestOfAuthorisation
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfBadHosts.html">		TestOfBadHosts
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfBrowserCookies.html">		TestOfBrowserCookies
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfBrowserFrames.html">		TestOfBrowserFrames
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfBrowserNavigation.html">		TestOfBrowserNavigation
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfCallingTearDownAfterExceptions.html">		TestOfCallingTearDownAfterExceptions
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfCallSchedule.html">		TestOfCallSchedule
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfCatchingExceptions.html">		TestOfCatchingExceptions
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfCheckbox.html">		TestOfCheckbox
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfCollector.html">		TestOfCollector
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfCommandLineParsing.html">		TestOfCommandLineParsing
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfCompatibility.html">		TestOfCompatibility
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfContext.html">		TestOfContext
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfCookie.html">		TestOfCookie
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfCookieJar.html">		TestOfCookieJar
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfDirectRoute.html">		TestOfDirectRoute
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfEclipse.html">		TestOfEclipse
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfEncodedParts.html">		TestOfEncodedParts
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfEncoding.html">		TestOfEncoding
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfEquality.html">		TestOfEquality
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfErrorQueue.html">		TestOfErrorQueue
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfErrors.html">		TestOfErrors
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfErrorsExcludingPHP52AndAbove.html">		TestOfErrorsExcludingPHP52AndAbove
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfErrorTrap.html">		TestOfErrorTrap
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfExceptionExpectation.html">		TestOfExceptionExpectation
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfExceptionThrownInSetUpDoesNotRunTestBody.html">		TestOfExceptionThrownInSetUpDoesNotRunTestBody
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfExceptionTrap.html">		TestOfExceptionTrap
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfExpectExceptionWithSetUp.html">		TestOfExpectExceptionWithSetUp
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfFetchingUrlParameters.html">		TestOfFetchingUrlParameters
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfFieldExpectation.html">		TestOfFieldExpectation
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfForm.html">		TestOfForm
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfFormHeaders.html">		TestOfFormHeaders
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfFormsCreatedFromEventStream.html">		TestOfFormsCreatedFromEventStream
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfFrameNavigation.html">		TestOfFrameNavigation
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfFrameset.html">		TestOfFrameset
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfFramesetPageInterface.html">		TestOfFramesetPageInterface
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfFrameUrl.html">		TestOfFrameUrl
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfGoogleVisualisation.html">		TestOfGoogleVisualisation
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfHeaderExpectations.html">		TestOfHeaderExpectations
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfHistory.html">		TestOfHistory
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfHtmlPage.html">		TestOfHtmlPage
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfHtmlSaxParser.html">		TestOfHtmlSaxParser
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfHttpHeaders.html">		TestOfHttpHeaders
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfHttpPostRequest.html">		TestOfHttpPostRequest
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfHttpRedirects.html">		TestOfHttpRedirects
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfHttpRequest.html">		TestOfHttpRequest
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfHttpResponse.html">		TestOfHttpResponse
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfIdentity.html">		TestOfIdentity
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfImplementations.html">		TestOfImplementations
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfInequality.html">		TestOfInequality
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfInterfaceMocksWithConstruct.html">		TestOfInterfaceMocksWithConstruct
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfInterfaceMocksWithHintInSignature.html">		TestOfInterfaceMocksWithHintInSignature
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfIsA.html">		TestOfIsA
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfLabelTag.html">		TestOfLabelTag
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfLexer.html">		TestOfLexer
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfLexerHandlers.html">		TestOfLexerHandlers
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfLexerModes.html">		TestOfLexerModes
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfLiveBrowser.html">		TestOfLiveBrowser
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfMethodExistence.html">		TestOfMethodExistence
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfMockComparisons.html">		TestOfMockComparisons
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfMockExpectations.html">		TestOfMockExpectations
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfMockExpectationsThatPass.html">		TestOfMockExpectationsThatPass
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfMockGeneration.html">		TestOfMockGeneration
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfMockingClassesWithStaticMethods.html">		TestOfMockingClassesWithStaticMethods
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfMockInterfaces.html">		TestOfMockInterfaces
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfMockObjectsOutput.html">		TestOfMockObjectsOutput
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfMockReturns.html">		TestOfMockReturns
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfNestingTags.html">		TestOfNestingTags
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfNonIdentity.html">		TestOfNonIdentity
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfNotA.html">		TestOfNotA
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfPageBuilder.html">		TestOfPageBuilder
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfPageHeaders.html">		TestOfPageHeaders
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfPageInterface.html">		TestOfPageInterface
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfPageParsing.html">		TestOfPageParsing
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfPageScraping.html">		TestOfPageScraping
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfParallelRegex.html">		TestOfParallelRegex
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfParametersExpectation.html">		TestOfParametersExpectation
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfParsedPageAccess.html">		TestOfParsedPageAccess
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfPartialMocks.html">		TestOfPartialMocks
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfPastBugs.html">		TestOfPastBugs
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfPatternCollector.html">		TestOfPatternCollector
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfPatterns.html">		TestOfPatterns
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfPHP4StyleSuperClassConstruct.html">		TestOfPHP4StyleSuperClassConstruct
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfPHP5AbstractMethodMocking.html">		TestOfPHP5AbstractMethodMocking
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfPHP5StaticMethodMocking.html">		TestOfPHP5StaticMethodMocking
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfPHP52RecoverableErrors.html">		TestOfPHP52RecoverableErrors
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfProxyRoute.html">		TestOfProxyRoute
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfRadioGroup.html">		TestOfRadioGroup
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfRealm.html">		TestOfRealm
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfReference.html">		TestOfReference
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfReflection.html">		TestOfReflection
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfReflectionWithTypeHints.html">		TestOfReflectionWithTypeHints
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfSampleInterfaceWithClone.html">		TestOfSampleInterfaceWithClone
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfSelection.html">		TestOfSelection
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfShell.html">		TestOfShell
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfShellTestCase.html">		TestOfShellTestCase
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfShellTesterAndShell.html">		TestOfShellTesterAndShell
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfSimpleHtmlLexer.html">		TestOfSimpleHtmlLexer
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfSimpleSignatureMap.html">		TestOfSimpleSignatureMap
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfSimpleStickyError.html">		TestOfSimpleStickyError
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfSkippingNoMatterWhat.html">		TestOfSkippingNoMatterWhat
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfSkippingOrElse.html">		TestOfSkippingOrElse
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfSkippingTwiceOver.html">		TestOfSkippingTwiceOver
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfSpecialMethods.html">		TestOfSpecialMethods
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfSpl.html">		TestOfSpl
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfStackTrace.html">		TestOfStackTrace
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfStateStack.html">		TestOfStateStack
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfTag.html">		TestOfTag
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfTagGroup.html">		TestOfTagGroup
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfTestDoxReporter.html">		TestOfTestDoxReporter
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfTextArea.html">		TestOfTextArea
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfTextExpectations.html">		TestOfTextExpectations
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfTextExtraction.html">		TestOfTextExtraction
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfTextFormatting.html">		TestOfTextFormatting
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfThrowingErrorsFromMocks.html">		TestOfThrowingErrorsFromMocks
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfThrowingExceptionsFromMocks.html">		TestOfThrowingExceptionsFromMocks
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfThrowingExceptionsInTearDown.html">		TestOfThrowingExceptionsInTearDown
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfUnitTester.html">		TestOfUnitTester
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfUploadWidget.html">		TestOfUploadWidget
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfUrl.html">		TestOfUrl
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfVisualShell.html">		TestOfVisualShell
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfWidget.html">		TestOfWidget
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfWithin.html">		TestOfWithin
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfXmlResultsParsing.html">		TestOfXmlResultsParsing
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestOfXmlStructureParsing.html">		TestOfXmlStructureParsing
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestRunnerForLeftOverAndNotEnoughErrors.html">		TestRunnerForLeftOverAndNotEnoughErrors
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestSuite.html">		TestSuite
		</a><br>
			<a href="../../SimpleTest/UnitTester/TestThatShouldNotBeSkipped.html">		TestThatShouldNotBeSkipped
		</a><br>
			<a href="../../SimpleTest/UnitTester/TextExpectation.html">		TextExpectation
		</a><br>
			<a href="../../SimpleTest/UnitTester/TextReporter.html">		TextReporter
		</a><br>
			<a href="../../SimpleTest/UnitTester/UnitTestCase.html">		UnitTestCase
		</a><br>
			<a href="../../SimpleTest/UnitTester/UnitTests.html">		UnitTests
		</a><br>
			<a href="../../SimpleTest/UnitTester/UnwantedPatternExpectation.html">		UnwantedPatternExpectation
		</a><br>
			<a href="../../SimpleTest/UnitTester/WantedPatternExpectation.html">		WantedPatternExpectation
		</a><br>
			<a href="../../SimpleTest/UnitTester/WithinMarginExpectation.html">		WithinMarginExpectation
		</a><br>
			<a href="../../SimpleTest/UnitTester/XmlReporter.html">		XmlReporter
		</a><br>
		<b>Extensions</b><br>			<a href="../../SimpleTest/Extensions/PHPUnit_TestCase.html">		PHPUnit_TestCase
		</a><br>
			<a href="../../SimpleTest/Extensions/TestCase.html">		TestCase
		</a><br>
			<a href="../../SimpleTest/Extensions/TestOfPearAdapter.html">		TestOfPearAdapter
		</a><br>
			<a href="../../SimpleTest/Extensions/TestOfPhpUnitAdapter.html">		TestOfPhpUnitAdapter
		</a><br>
					<a href="../../SimpleTest/SimpleTestCompatibility.html">		SimpleTestCompatibility
		</a><br>
		<b>Eclipse</b><br>			<a href="../../SimpleTest/Eclipse/EclipseInvoker.html">		EclipseInvoker
		</a><br>
			<a href="../../SimpleTest/Eclipse/EclipseReporter.html">		EclipseReporter
		</a><br>
		<b>MockObjects</b><br>			<a href="../../SimpleTest/MockObjects/AnythingExpectation.html">		AnythingExpectation
		</a><br>
			<a href="../../SimpleTest/MockObjects/CallCountExpectation.html">		CallCountExpectation
		</a><br>
			<a href="../../SimpleTest/MockObjects/FailedExpectation.html">		FailedExpectation
		</a><br>
			<a href="../../SimpleTest/MockObjects/FalseExpectation.html">		FalseExpectation
		</a><br>
			<a href="../../SimpleTest/MockObjects/MaximumCallCountExpectation.html">		MaximumCallCountExpectation
		</a><br>
			<a href="../../SimpleTest/MockObjects/MinimumCallCountExpectation.html">		MinimumCallCountExpectation
		</a><br>
			<a href="../../SimpleTest/MockObjects/Mock.html">		Mock
		</a><br>
			<a href="../../SimpleTest/MockObjects/MockGenerator.html">		MockGenerator
		</a><br>
			<a href="../../SimpleTest/MockObjects/MockWithInjectedTestCase.html">		MockWithInjectedTestCase
		</a><br>
			<a href="../../SimpleTest/MockObjects/ParametersExpectation.html">		ParametersExpectation
		</a><br>
			<a href="../../SimpleTest/MockObjects/SimpleByReference.html">		SimpleByReference
		</a><br>
			<a href="../../SimpleTest/MockObjects/SimpleByValue.html">		SimpleByValue
		</a><br>
			<a href="../../SimpleTest/MockObjects/SimpleCallSchedule.html">		SimpleCallSchedule
		</a><br>
			<a href="../../SimpleTest/MockObjects/SimpleErrorThrower.html">		SimpleErrorThrower
		</a><br>
			<a href="../../SimpleTest/MockObjects/SimpleMock.html">		SimpleMock
		</a><br>
			<a href="../../SimpleTest/MockObjects/SimpleSignatureMap.html">		SimpleSignatureMap
		</a><br>
			<a href="../../SimpleTest/MockObjects/SimpleThrower.html">		SimpleThrower
		</a><br>
			<a href="../../SimpleTest/MockObjects/Stub.html">		Stub
		</a><br>
			<a href="../../SimpleTest/MockObjects/TrueExpectation.html">		TrueExpectation
		</a><br>
				</div>
		</div>
	</div>

<div id="body">
	    		<h1>Class: TestRadioFields</h1>
	<p style="margin: 0px;">Source Location: /simpletest/test/acceptance_test.php</p>
	

<div class="leftcol">
	<h3><a href="#class_details">Class Overview</a> <span class="smalllinenumber">[line 106]</span></h3>
	<div id="classTree"><pre><a href="../../SimpleTest/UnitTester/SimpleTestCase.html">SimpleTestCase</a>
   |
   --<a href="../../SimpleTest/WebTester/WebTestCase.html">WebTestCase</a>
      |
      --<a href="../../SimpleTest/WebTester/SimpleTestAcceptanceTest.html">SimpleTestAcceptanceTest</a>
         |
         --TestRadioFields</pre>
</div>
	<div class="small">
	<p>Test case for testing of web pages. Allows     fetching of pages, parsing of HTML and     submitting forms.</p>
		<h4>Author(s):</h4>
	<ul>
			</ul>
	<h4>Version:</h4>
	<ul>
			</ul>

	<h4>Copyright:</h4>
	<ul>
			</li>
	</div>
</div>

<div class="middlecol">
	<h3><a href="#class_vars">Variables</a></h3>
	<ul class="small">
			</ul>
	<h3><a href="#class_consts">Constants</a></h3>
	<ul class="small">
			</ul>
</div>
<div class="rightcol">
	<h3><a href="#class_methods">Methods</a></h3>
	<ul class="small">
				<li><a href="../../SimpleTest/WebTester/TestRadioFields.html#methodtestSetFieldAsInteger">testSetFieldAsInteger</a></li>
				<li><a href="../../SimpleTest/WebTester/TestRadioFields.html#methodtestSetFieldAsString">testSetFieldAsString</a></li>
			</ul>
</div>

<div id="content">
<hr>
	<div class="contents">
	</div>

	<div class="leftCol">
    	<h2>Inherited Variables</h2>
			<div class="indent">
		<h3>Class: <a href="../../SimpleTest/WebTester/WebTestCase.html">WebTestCase</a></h3>
		<div class="small">
			<dl>
						<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#var$_browser">WebTestCase::$_browser</a>
			</dt>
			<dd>
				 
			</dd>
						<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#var$_ignore_errors">WebTestCase::$_ignore_errors</a>
			</dt>
			<dd>
				 
			</dd>
						</dl>
		</div>
		</div>
			<div class="indent">
		<h3>Class: <a href="../../SimpleTest/UnitTester/SimpleTestCase.html">SimpleTestCase</a></h3>
		<div class="small">
			<dl>
						<dt>
				<a href="../../SimpleTest/UnitTester/SimpleTestCase.html#var$_label">SimpleTestCase::$_label</a>
			</dt>
			<dd>
				 
			</dd>
						<dt>
				<a href="../../SimpleTest/UnitTester/SimpleTestCase.html#var$_observers">SimpleTestCase::$_observers</a>
			</dt>
			<dd>
				 
			</dd>
						<dt>
				<a href="../../SimpleTest/UnitTester/SimpleTestCase.html#var$_reporter">SimpleTestCase::$_reporter</a>
			</dt>
			<dd>
				 
			</dd>
						<dt>
				<a href="../../SimpleTest/UnitTester/SimpleTestCase.html#var$_should_skip">SimpleTestCase::$_should_skip</a>
			</dt>
			<dd>
				 
			</dd>
						</dl>
		</div>
		</div>
		<h2>Inherited Constants</h2>
		</div>

	<div class="rightCol">
	<h2>Inherited Methods</h2>
			<div class="indent">
		<h3>Class: <a href="../../SimpleTest/WebTester/SimpleTestAcceptanceTest.html">SimpleTestAcceptanceTest</a></h3>
		<dl class="small">
						<dt>
				<a href="../../SimpleTest/WebTester/SimpleTestAcceptanceTest.html#methodsamples">SimpleTestAcceptanceTest::samples()</a>
			</dt>
			<dd>
				
			</dd>
				</dl>
		</div>
			<div class="indent">
		<h3>Class: <a href="../../SimpleTest/WebTester/WebTestCase.html">WebTestCase</a></h3>
		<dl class="small">
						<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodWebTestCase">WebTestCase::WebTestCase()</a>
			</dt>
			<dd>
				Creates an empty test case. Should be subclassed     with test methods for a functional test case.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodaddHeader">WebTestCase::addHeader()</a>
			</dt>
			<dd>
				Adds a header to every fetch.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodafter">WebTestCase::after()</a>
			</dt>
			<dd>
				Announces the end of the test. Includes private clean up.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodageCookies">WebTestCase::ageCookies()</a>
			</dt>
			<dd>
				Moves cookie expiry times back into the past.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertAuthentication">WebTestCase::assertAuthentication()</a>
			</dt>
			<dd>
				Attempt to match the authentication type within     the security realm we are currently matching.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertClickable">WebTestCase::assertClickable()</a>
			</dt>
			<dd>
				Checks for a click target.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertCookie">WebTestCase::assertCookie()</a>
			</dt>
			<dd>
				Checks that a cookie is set for the current page     and optionally checks the value.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertEqual">WebTestCase::assertEqual()</a>
			</dt>
			<dd>
				Will trigger a pass if the two parameters have     the same value only. Otherwise a fail. This     is for testing hand extracted text, etc.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertFalse">WebTestCase::assertFalse()</a>
			</dt>
			<dd>
				Will be true on false and vice versa. False     is the PHP definition of false, so that null,     empty strings, zero and an empty array all count     as false.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertField">WebTestCase::assertField()</a>
			</dt>
			<dd>
				Confirms that the form element is currently set     to the expected value. A missing form will always     fail. If no value is given then only the existence     of the field is checked.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertFieldById">WebTestCase::assertFieldById()</a>
			</dt>
			<dd>
				Confirms that the form element is currently set     to the expected value. A missing form will always     fail. If no ID is given then only the existence     of the field is checked.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertFieldByName">WebTestCase::assertFieldByName()</a>
			</dt>
			<dd>
				Confirms that the form element is currently set     to the expected value. A missing form element will always     fail. If no value is given then only the existence     of the field is checked.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertHeader">WebTestCase::assertHeader()</a>
			</dt>
			<dd>
				Checks each header line for the required value. If no     value is given then only an existence check is made.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertHeaderPattern">WebTestCase::assertHeaderPattern()</a>
			</dt>
			<dd>
				
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertImage">WebTestCase::assertImage()</a>
			</dt>
			<dd>
				Checks for a valid image with atht alt text or title.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertLink">WebTestCase::assertLink()</a>
			</dt>
			<dd>
				Tests for the presence of a link label. Match is     case insensitive with normalised space.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertLinkById">WebTestCase::assertLinkById()</a>
			</dt>
			<dd>
				Tests for the presence of a link id attribute.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertMime">WebTestCase::assertMime()</a>
			</dt>
			<dd>
				Checks the mime type against a list     of possible values.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertNoAuthentication">WebTestCase::assertNoAuthentication()</a>
			</dt>
			<dd>
				Checks that no authentication is necessary to view     the desired page.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertNoCookie">WebTestCase::assertNoCookie()</a>
			</dt>
			<dd>
				Checks that no cookie is present or that it has     been successfully cleared.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertNoHeader">WebTestCase::assertNoHeader()</a>
			</dt>
			<dd>
				Confirms that the header type has not been received.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertNoLink">WebTestCase::assertNoLink()</a>
			</dt>
			<dd>
				Tests for the non-presence of a link label. Match is     case insensitive with normalised space.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertNoLinkById">WebTestCase::assertNoLinkById()</a>
			</dt>
			<dd>
				Tests for the non-presence of a link label. Match is     case insensitive with normalised space.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertNoPattern">WebTestCase::assertNoPattern()</a>
			</dt>
			<dd>
				Will trigger a pass if the perl regex pattern     is not present in raw content.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertNotEqual">WebTestCase::assertNotEqual()</a>
			</dt>
			<dd>
				Will trigger a pass if the two parameters have     a different value. Otherwise a fail. This     is for testing hand extracted text, etc.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertNoText">WebTestCase::assertNoText()</a>
			</dt>
			<dd>
				Will trigger a pass if the text is not found in the plain     text form of the page.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertNoUnwantedHeader">WebTestCase::assertNoUnwantedHeader()</a>
			</dt>
			<dd>
				
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertNoUnwantedPattern">WebTestCase::assertNoUnwantedPattern()</a>
			</dt>
			<dd>
				
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertNoUnwantedText">WebTestCase::assertNoUnwantedText()</a>
			</dt>
			<dd>
				
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertPattern">WebTestCase::assertPattern()</a>
			</dt>
			<dd>
				Will trigger a pass if the Perl regex pattern     is found in the raw content.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertRealm">WebTestCase::assertRealm()</a>
			</dt>
			<dd>
				Attempts to match the current security realm.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertResponse">WebTestCase::assertResponse()</a>
			</dt>
			<dd>
				Checks the response code against a list     of possible values.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertSubmit">WebTestCase::assertSubmit()</a>
			</dt>
			<dd>
				Checks for a valid button label.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertText">WebTestCase::assertText()</a>
			</dt>
			<dd>
				Will trigger a pass if the text is found in the plain     text form of the page.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertTitle">WebTestCase::assertTitle()</a>
			</dt>
			<dd>
				Tests the text between the title tags.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertTrue">WebTestCase::assertTrue()</a>
			</dt>
			<dd>
				Called from within the test methods to register     passes and failures.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertWantedPattern">WebTestCase::assertWantedPattern()</a>
			</dt>
			<dd>
				
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodassertWantedText">WebTestCase::assertWantedText()</a>
			</dt>
			<dd>
				
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodauthenticate">WebTestCase::authenticate()</a>
			</dt>
			<dd>
				Retries a request after setting the authentication     for the current realm.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodback">WebTestCase::back()</a>
			</dt>
			<dd>
				Equivalent to hitting the back button on the     browser.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodbefore">WebTestCase::before()</a>
			</dt>
			<dd>
				Announces the start of the test.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodclearFrameFocus">WebTestCase::clearFrameFocus()</a>
			</dt>
			<dd>
				Clears the frame focus. All frames will be searched     for content.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodclick">WebTestCase::click()</a>
			</dt>
			<dd>
				Clicks a visible text item. Will first try buttons,     then links and then images.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodclickImage">WebTestCase::clickImage()</a>
			</dt>
			<dd>
				Clicks the submit image by some kind of label. Usually
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodclickImageById">WebTestCase::clickImageById()</a>
			</dt>
			<dd>
				Clicks the submit image by ID attribute. The owning     form will be submitted by this. Clicking outside of     the boundary of the coordinates will result in     a failure.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodclickImageByName">WebTestCase::clickImageByName()</a>
			</dt>
			<dd>
				Clicks the submit image by the name. Usually
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodclickLink">WebTestCase::clickLink()</a>
			</dt>
			<dd>
				Follows a link by name. Will click the first link     found with this link text by default, or a later     one if an index is given. Match is case insensitive     with normalised space.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodclickLinkById">WebTestCase::clickLinkById()</a>
			</dt>
			<dd>
				Follows a link by id attribute.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodclickSubmit">WebTestCase::clickSubmit()</a>
			</dt>
			<dd>
				Clicks the submit button by label. The owning     form will be submitted by this.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodclickSubmitById">WebTestCase::clickSubmitById()</a>
			</dt>
			<dd>
				Clicks the submit button by ID attribute. The owning     form will be submitted by this.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodclickSubmitByName">WebTestCase::clickSubmitByName()</a>
			</dt>
			<dd>
				Clicks the submit button by name attribute. The owning     form will be submitted by this.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodcreateBrowser">WebTestCase::createBrowser()</a>
			</dt>
			<dd>
				Creates a new default web browser object.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodforward">WebTestCase::forward()</a>
			</dt>
			<dd>
				Equivalent to hitting the forward button on the     browser.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodget">WebTestCase::get()</a>
			</dt>
			<dd>
				Fetches a page into the page buffer. If     there is no base for the URL then the     current base URL is used. After the fetch     the base URL reflects the new location.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodgetAssertionLine">WebTestCase::getAssertionLine()</a>
			</dt>
			<dd>
				Uses a stack trace to find the line of an assertion.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodgetBrowser">WebTestCase::getBrowser()</a>
			</dt>
			<dd>
				Gets a current browser reference for setting     special expectations or for detailed     examination of page fetches.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodgetCookie">WebTestCase::getCookie()</a>
			</dt>
			<dd>
				Gets the cookie value for the current browser context.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodgetFrameFocus">WebTestCase::getFrameFocus()</a>
			</dt>
			<dd>
				Accessor for current frame focus. Will be     false if no frame has focus.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodgetTransportError">WebTestCase::getTransportError()</a>
			</dt>
			<dd>
				Gets the last response error.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodgetUrl">WebTestCase::getUrl()</a>
			</dt>
			<dd>
				Accessor for the currently selected URL.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodhead">WebTestCase::head()</a>
			</dt>
			<dd>
				Does a HTTP HEAD fetch, fetching only the page     headers. The current base URL is unchanged by this.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodignoreCookies">WebTestCase::ignoreCookies()</a>
			</dt>
			<dd>
				Switches off cookie sending and recieving.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodignoreErrors">WebTestCase::ignoreErrors()</a>
			</dt>
			<dd>
				Skips errors for the next request only. You might     want to confirm that a page is unreachable for     example.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodignoreFrames">WebTestCase::ignoreFrames()</a>
			</dt>
			<dd>
				Disables frames support. Frames will not be fetched     and the frameset page will be used instead.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodpost">WebTestCase::post()</a>
			</dt>
			<dd>
				Fetches a page by POST into the page buffer.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodrestart">WebTestCase::restart()</a>
			</dt>
			<dd>
				Simulates the closing and reopening of the browser.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodretry">WebTestCase::retry()</a>
			</dt>
			<dd>
				Equivalent to hitting the retry button on the     browser. Will attempt to repeat the page fetch.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodsetBrowser">WebTestCase::setBrowser()</a>
			</dt>
			<dd>
				Gets a current browser reference for setting     special expectations or for detailed     examination of page fetches.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodsetConnectionTimeout">WebTestCase::setConnectionTimeout()</a>
			</dt>
			<dd>
				Sets the socket timeout for opening a connection and     receiving at least one byte of information.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodsetCookie">WebTestCase::setCookie()</a>
			</dt>
			<dd>
				Sets a cookie in the current browser.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodsetField">WebTestCase::setField()</a>
			</dt>
			<dd>
				Sets all form fields with that label, or name if there     is no label attached.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodsetFieldById">WebTestCase::setFieldById()</a>
			</dt>
			<dd>
				Sets all form fields with that id.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodsetFieldByName">WebTestCase::setFieldByName()</a>
			</dt>
			<dd>
				Sets all form fields with that name.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodsetFrameFocus">WebTestCase::setFrameFocus()</a>
			</dt>
			<dd>
				Sets the focus by name.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodsetFrameFocusByIndex">WebTestCase::setFrameFocusByIndex()</a>
			</dt>
			<dd>
				Sets the focus by index. The integer index starts from 1.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodsetMaximumRedirects">WebTestCase::setMaximumRedirects()</a>
			</dt>
			<dd>
				Sets the maximum number of redirects before     the web page is loaded regardless.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodshowHeaders">WebTestCase::showHeaders()</a>
			</dt>
			<dd>
				Dumps the current HTTP headers for debugging.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodshowRequest">WebTestCase::showRequest()</a>
			</dt>
			<dd>
				Dumps the current request for debugging.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodshowSource">WebTestCase::showSource()</a>
			</dt>
			<dd>
				Dumps the current HTML source for debugging.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodshowText">WebTestCase::showText()</a>
			</dt>
			<dd>
				Dumps the visible text only for debugging.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodsubmitFormById">WebTestCase::submitFormById()</a>
			</dt>
			<dd>
				Submits a form by the ID.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methodunsetBrowser">WebTestCase::unsetBrowser()</a>
			</dt>
			<dd>
				Clears the current browser reference to help the     PHP garbage collector.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#methoduseProxy">WebTestCase::useProxy()</a>
			</dt>
			<dd>
				Sets proxy to use on all requests for when     testing from behind a firewall. Set URL     to false to disable.
			</dd>
					<dt>
				<a href="../../SimpleTest/WebTester/WebTestCase.html#method_assertFieldValue">WebTestCase::_assertFieldValue()</a>
			</dt>
			<dd>
				Tests the field value against the expectation.
			</dd>
				</dl>
		</div>
			<div class="indent">
		<h3>Class: <a href="../../SimpleTest/UnitTester/SimpleTestCase.html">SimpleTestCase</a></h3>
		<dl class="small">
						<dt>
				<a href="../../SimpleTest/UnitTester/SimpleTestCase.html#methodSimpleTestCase">SimpleTestCase::SimpleTestCase()</a>
			</dt>
			<dd>
				Sets up the test with no display.
			</dd>
					<dt>
				<a href="../../SimpleTest/UnitTester/SimpleTestCase.html#methodafter">SimpleTestCase::after()</a>
			</dt>
			<dd>
				Announces the end of the test. Includes private clean up.
			</dd>
					<dt>
				<a href="../../SimpleTest/UnitTester/SimpleTestCase.html#methodassert">SimpleTestCase::assert()</a>
			</dt>
			<dd>
				Runs an expectation directly, for extending the     tests with new expectation classes.
			</dd>
					<dt>
				<a href="../../SimpleTest/UnitTester/SimpleTestCase.html#methodassertExpectation">SimpleTestCase::assertExpectation()</a>
			</dt>
			<dd>
				
			</dd>
					<dt>
				<a href="../../SimpleTest/UnitTester/SimpleTestCase.html#methodbefore">SimpleTestCase::before()</a>
			</dt>
			<dd>
				Announces the start of the test.
			</dd>
					<dt>
				<a href="../../SimpleTest/UnitTester/SimpleTestCase.html#methodcreateInvoker">SimpleTestCase::createInvoker()</a>
			</dt>
			<dd>
				Used to invoke the single tests.
			</dd>
					<dt>
				<a href="../../SimpleTest/UnitTester/SimpleTestCase.html#methoddump">SimpleTestCase::dump()</a>
			</dt>
			<dd>
				Sends a formatted dump of a variable to the     test suite for those emergency debugging     situations.
			</dd>
					<dt>
				<a href="../../SimpleTest/UnitTester/SimpleTestCase.html#methoderror">SimpleTestCase::error()</a>
			</dt>
			<dd>
				Formats a PHP error and dispatches it to the     reporter.
			</dd>
					<dt>
				<a href="../../SimpleTest/UnitTester/SimpleTestCase.html#methodexception">SimpleTestCase::exception()</a>
			</dt>
			<dd>
				Formats an exception and dispatches it to the     reporter.
			</dd>
					<dt>
				<a href="../../SimpleTest/UnitTester/SimpleTestCase.html#methodfail">SimpleTestCase::fail()</a>
			</dt>
			<dd>
				Sends a fail event with a message.
			</dd>
					<dt>
				<a href="../../SimpleTest/UnitTester/SimpleTestCase.html#methodgetAssertionLine">SimpleTestCase::getAssertionLine()</a>
			</dt>
			<dd>
				Uses a stack trace to find the line of an assertion.
			</dd>
					<dt>
				<a href="../../SimpleTest/UnitTester/SimpleTestCase.html#methodgetLabel">SimpleTestCase::getLabel()</a>
			</dt>
			<dd>
				Accessor for the test name for subclasses.
			</dd>
					<dt>
				<a href="../../SimpleTest/UnitTester/SimpleTestCase.html#methodgetSize">SimpleTestCase::getSize()</a>
			</dt>
			<dd>
				Accessor for the number of subtests including myelf.
			</dd>
					<dt>
				<a href="../../SimpleTest/UnitTester/SimpleTestCase.html#methodgetTests">SimpleTestCase::getTests()</a>
			</dt>
			<dd>
				Gets a list of test names. Normally that will     be all internal methods that start with the     name &quot;test&quot;. This method should be overridden     if you want a different rule.
			</dd>
					<dt>
				<a href="../../SimpleTest/UnitTester/SimpleTestCase.html#methodpass">SimpleTestCase::pass()</a>
			</dt>
			<dd>
				
			</dd>
					<dt>
				<a href="../../SimpleTest/UnitTester/SimpleTestCase.html#methodrun">SimpleTestCase::run()</a>
			</dt>
			<dd>
				Uses reflection to run every method within itself     starting with the string &quot;test&quot; unless a method     is specified.
			</dd>
					<dt>
				<a href="../../SimpleTest/UnitTester/SimpleTestCase.html#methodsendMessage">SimpleTestCase::sendMessage()</a>
			</dt>
			<dd>
				
			</dd>
					<dt>
				<a href="../../SimpleTest/UnitTester/SimpleTestCase.html#methodsetUp">SimpleTestCase::setUp()</a>
			</dt>
			<dd>
				Sets up unit test wide variables at the start     of each test method. To be overridden in     actual user test cases.
			</dd>
					<dt>
				<a href="../../SimpleTest/UnitTester/SimpleTestCase.html#methodsignal">SimpleTestCase::signal()</a>
			</dt>
			<dd>
				
			</dd>
					<dt>
				<a href="../../SimpleTest/UnitTester/SimpleTestCase.html#methodskip">SimpleTestCase::skip()</a>
			</dt>
			<dd>
				This is a placeholder for skipping tests. In this     method you place skipIf() and skipUnless() calls to     set the skipping state.
			</dd>
					<dt>
				<a href="../../SimpleTest/UnitTester/SimpleTestCase.html#methodskipIf">SimpleTestCase::skipIf()</a>
			</dt>
			<dd>
				Will issue a message to the reporter and tell the test     case to skip if the incoming flag is true.
			</dd>
					<dt>
				<a href="../../SimpleTest/UnitTester/SimpleTestCase.html#methodskipUnless">SimpleTestCase::skipUnless()</a>
			</dt>
			<dd>
				Will issue a message to the reporter and tell the test     case to skip if the incoming flag is false.
			</dd>
					<dt>
				<a href="../../SimpleTest/UnitTester/SimpleTestCase.html#methodtearDown">SimpleTestCase::tearDown()</a>
			</dt>
			<dd>
				Clears the data set in the setUp() method call.
			</dd>
					<dt>
				<a href="../../SimpleTest/UnitTester/SimpleTestCase.html#methodtell">SimpleTestCase::tell()</a>
			</dt>
			<dd>
				Sets up an observer for the test end.
			</dd>
					<dt>
				<a href="../../SimpleTest/UnitTester/SimpleTestCase.html#method_isTest">SimpleTestCase::_isTest()</a>
			</dt>
			<dd>
				Tests to see if the method is a test that should     be run. Currently any method that starts with 'test'     is a candidate unless it is the constructor.
			</dd>
				</dl>
		</div>
		</div>
	<br clear="all">
	<hr>

	<a name="class_details"></a>
	<h2>Class Details</h2>
	<p align="center"><strong>Test case for testing of web pages. Allows     fetching of pages, parsing of HTML and     submitting forms.
</strong></p>
	<p class="small" style="color: #334B66;">[ <a href="#top">Top</a> ]</p>

	<hr>
	<a name="class_vars"></a>
	<h2>Class Variables</h2>
	
	<hr>
	<a name="class_methods"></a>
	<h2>Class Methods</h2>
	
	<a name="methodtestSetFieldAsInteger"></a>
	<p></p>
	<h3>testSetFieldAsInteger</h3>
	<div class="indent">
		<p>
		<code>void testSetFieldAsInteger(
)</code>
		</p>
	
		<p class="linenumber">[line 107]</p>
				
	
	<h4>Parameters:</h4>
	<ul>
		</ul>
	</div>
	<p class="top">[ <a href="#top">Top</a> ]</p>
	<a name="methodtestSetFieldAsString"></a>
	<p></p>
	<h3>testSetFieldAsString</h3>
	<div class="indent">
		<p>
		<code>void testSetFieldAsString(
)</code>
		</p>
	
		<p class="linenumber">[line 114]</p>
				
	
	<h4>Parameters:</h4>
	<ul>
		</ul>
	</div>
	<p class="top">[ <a href="#top">Top</a> ]</p>

	<hr>
	<a name="class_consts"></a>
	<h2>Class Constants</h2>
	</div>
	<div id="credit">
		<hr>
		Documentation generated on Tue, 26 May 2009 19:19:09 +0200 by <a href="http://www.phpdoc.org">phpDocumentor 1.4.1</a>
	</div>
</div>
</body>
</html>
